requirements.txt
deep-translator==1.11.4
pytest==9.0.2
pytest-xdist==3.8.0
python-dotenv==1.2.1  
requests==2.32.5
selenium==4.40.0
webdriver-manager==4.0.2